pyproject.toml
[project]
name = "das-port-forwarder"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "paramiko>=3.5.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "paramiko>=3.5.1",
    "pyinstaller>=6.12.0",
    "python-dotenv>=1.0.1",
]
